refactor: SettingPage에서 setting하는 부분 훅으로 분리
