+++ frontend/src/pages/HumanChatPage/SettingPage.tsx
@@ -1,15 +1,12 @@
-import { ChangeEvent, useEffect } from 'react';
+import { useEffect } from 'react';
 import { useNavigate, useOutletContext } from 'react-router-dom';
 
 import ProfileSetting from '@components/ProfileSetting';
 
 import { HumanSocketManager } from '@business/services/SocketManager';
 
-import { useProfileInfo } from '@stores/zustandStores/useProfileInfo';
-
-import { arrayBuffer2Array } from '@utils/array';
-
 import type { OutletContext } from './HumanChatPage';
+import { useSettingPageProfileNicknameSetting } from './useSettingPageProfileNicknameSetting';
 
 export default function ChattingPage() {
   const socketManager = new HumanSocketManager();
@@ -18,19 +15,16 @@ export default function ChattingPage() {
 
   const {
     localVideoRef,
-    toggleVideo,
-    toggleAudio,
     cameraOptions,
     audioOptions,
+    profileChannel,
+    nicknameChannel,
+    toggleVideo,
+    toggleAudio,
     changeMyVideoTrack,
     changeMyAudioTrack,
     getMedia,
-    profileChannel,
-    nicknameChannel,
   }: OutletContext = useOutletContext();
-
-  const camList = cameraOptions.map(({ deviceId, label }) => ({ label, value: deviceId }));
-  const micList = audioOptions.map(({ deviceId, label }) => ({ label, value: deviceId }));
 
   useEffect(() => {
     if (!socketManager.connected) {
@@ -39,42 +33,11 @@ export default function ChattingPage() {
     getMedia({});
   }, []);
 
-  const { myNickname, myProfile, setMyNickname, setMyProfileImage } = useProfileInfo(state => ({
-    setMyNickname: state.setMyNickname,
-    setMyProfileImage: state.setMyProfile,
-    myNickname: state.myNickname,
-    myProfile: state.myProfile,
-  }));
+  const camList = cameraOptions.map(({ deviceId, label }) => ({ label, value: deviceId }));
+  const micList = audioOptions.map(({ deviceId, label }) => ({ label, value: deviceId }));
 
-  const setLocalProfileImage = async (e: ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) {
-      return;
-    }
-    const { type } = file;
-
-    const arrayBuffer = await file.arrayBuffer();
-    setMyProfileImage({ arrayBuffer, type });
-  };
-
-  const setLocalNickname = (e: ChangeEvent<HTMLInputElement>) => {
-    setMyNickname(e.target.value);
-  };
-
-  const sendProfileInfoWithNavigate = () => {
-    if (profileChannel.current?.readyState === 'open' && myProfile) {
-      const dataArray = arrayBuffer2Array(myProfile.arrayBuffer);
-      const sendJson = JSON.stringify({ arrayBuffer: dataArray, type: myProfile.type });
-
-      profileChannel.current?.send?.(sendJson);
-    }
-
-    if (nicknameChannel.current?.readyState === 'open' && myNickname) {
-      nicknameChannel.current?.send?.(myNickname);
-    }
-
-    navigate('..');
-  };
+  const { setLocalNickname, setLocalProfileImage, sendProfileInfoWithNavigateBefore } =
+    useSettingPageProfileNicknameSetting({ profileChannel, nicknameChannel });
 
   return (
     <ProfileSetting
@@ -85,7 +48,7 @@ export default function ChattingPage() {
       camList={camList}
       micList={micList}
       videoRef={localVideoRef}
-      onConfirm={sendProfileInfoWithNavigate}
+      onConfirm={sendProfileInfoWithNavigateBefore}
       onChangeProfileImage={setLocalProfileImage}
       onChangeNickname={setLocalNickname}
     />
